requirements.txt
biopython==1.70
numpy==1.14.0
pandas==0.22.0
primer3-py==0.5.5
pysam==0.13
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
numpydoc==0.7.0
